fix: Change all tool file names to kebab format
