+++ web/src/pages/add-knowledge/components/knowledge-file/hooks.ts
@@ -14,7 +14,7 @@ import { useOneNamespaceEffectsLoading } from '@/hooks/store-hooks';
 import { useFetchTenantInfo } from '@/hooks/user-setting-hooks';
 import { Pagination } from '@/interfaces/common';
 import { IChangeParserConfigRequestBody } from '@/interfaces/request/document';
-import { getUnSupportedFilesCount } from '@/utils/documentUtils';
+import { getUnSupportedFilesCount } from '@/utils/document-util';
 import { PaginationProps, UploadFile } from 'antd';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useDispatch, useNavigate, useSelector } from 'umi';
