+++ web/src/pages/add-knowledge/components/knowledge-file/parsing-action-cell/index.tsx
@@ -2,7 +2,7 @@ import { useShowDeleteConfirm, useTranslate } from '@/hooks/common-hooks';
 import { useRemoveDocument } from '@/hooks/document-hooks';
 import { IKnowledgeFile } from '@/interfaces/database/knowledge';
 import { api_host } from '@/utils/api';
-import { downloadFile } from '@/utils/fileUtil';
+import { downloadFile } from '@/utils/file-util';
 import {
   DeleteOutlined,
   DownloadOutlined,
