+++ web/src/pages/file-manager/index.tsx
@@ -18,8 +18,8 @@ import FileUploadModal from '@/components/file-upload-modal';
 import RenameModal from '@/components/rename-modal';
 import SvgIcon from '@/components/svg-icon';
 import { useTranslate } from '@/hooks/common-hooks';
-import { formatNumberWithThousandsSeparator } from '@/utils/commonUtil';
-import { getExtension } from '@/utils/documentUtils';
+import { formatNumberWithThousandsSeparator } from '@/utils/common-util';
+import { getExtension } from '@/utils/document-util';
 import ConnectToKnowledgeModal from './connect-to-knowledge-modal';
 import FolderCreateModal from './folder-create-modal';
 import styles from './index.less';
