+++ web/src/utils/common-util.ts
