+++ web/src/utils/document-util.ts
